ANDROID: ABI fixup for abi break in struct dst_ops
In commit 079d4f3ff06b ("BACKPORT: net: fix __dst_negative_advice() race") the
struct dst_ops callback negative_advice is callback changes function
parameters.  But as this pointer is part of a structure that is tracked
in the ABI checker, the tool triggers when this is changed.

However, the callback pointer is internal to the networking stack, so
changing the function type is safe, so needing to preserve this is not
required.  To do so, switch the function pointer type back to the old
one so that the checking tools pass, AND then do a hard cast of the
function pointer to the new type when assigning and calling the
function.

Bug: 343727534
Fixes: 079d4f3ff06b ("BACKPORT: net: fix __dst_negative_advice() race")
Change-Id: I48d4ab4bbd29f8edc8fbd7923828b7f78a23e12e
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
(cherry picked from commit ff29a6cf6e84f4ccbab2d0853cf20e6c2821ef38)
Signed-off-by: mrsrimar22 <mrsrimar22@gmail.com>